remove gensym function to use object() instead
I had written a gensym function that needed to be threaded throughout any
environment needing gensyms in order to avoid conflicts. It turns out that
objects are guaranteed unique IDs so long as they don't overlap, so creating a
base object to serve as the identity checkers is a better option.