typecheck TRSs in AddRuleProposer for "operator invention"
If we add a bunch of operators to the typesystem whose types are type variables
rather than type operators or type bindings, then we have operators of unknown
type whose type is specified by its usage in the learned TRS. In order to
support that, we need to make sure the operators are being used in a type
consistent way, which means we need to typecheck the TRS each time we propose to
add something to it. These changes support that sort of typechecking.